Fix a bug where sometimes covers wouldn't download correctly
In particular, if the URL didn't have a file extension, the code would
guess a completely nonsensical name for the cover path, e.g. if the URL
was

    https://example.org/images/books/cover-of-this-book

then it would try to download the image with the name

    org/images/books/cover-of-this-book

which caused a "file not found error" because there's no `org/images/books`
directory to download the image into.

This patch modifies the download code so it uses the HTTP Content-Type
header to decide what file extension to use, rather than the URL.

It's probably still a bit flaky, but it fixes the immediate problem I
was having tonight and should be a better foundation for future work.