Make it faster to retrieve header images from cache
This turns out to be the slow part of rebuilding the site from scratch --
even though we were saving the PNG file, we were then reopening it and
creating the data URL on every build.  This is unnecessary!

Just save the final data URL and skip reopening a PNG each time.