Fix a bug where placeholder image blocks were the wrong size
When you loaded the page on mobile, the solid colour rectangles that
appear before images load would appear… but in the wrong size.

The problem is that:

* By default, `.bk_t.ar_p` (portrait thumbnail wrapper) has `height: 130px`
* On mobile, the `<img>` is made slightly bigger with `height: 150px`

But I wasn't updating the `.bk_t.ar_p`, so the image was overflowing its
container and the placeholder block of colour.  Making the heights
consistent should fix that.